refactor (maintenance): change namespace
Move the class from its own namespace to the Foundation namespace.

Relates to #36474
